Update Readme.md
I added the markdown format. I have decided to let the script author select a title for the script. Also fill in the dependencies that are needed and explain how to obtain the key of google map api. You can complete the how to use it if you want with the options. :)